+++ companion-module-streamline-techhub/src/main.js
@@ -1,0 +1,4 @@
+'use strict';
+const { TechHubInstance } = require('./instance');
+// Companion API 2 loads the default module export.
+module.exports = TechHubInstance;